Move in_workspace_with_projects_file to tests/util/mod.rs
