+++ tests/clone.rs
@@ -5,21 +5,17 @@ extern crate tempdir;
 mod util;
 
 use std::collections::HashSet;
-use std::fs::create_dir_all;
 use std::hash::Hash;
-use std::path::Path;
 
 use git2::Repository;
-use tempdir::TempDir;
 
 use gws::color::palette::Palette;
 use gws::commands::clone::Clone;
 use gws::commands::common::DirectoryCommand;
 use gws::config::data::Workspace;
-use gws::config::read::read_workspace_file;
 
 use util::in_example_workspace;
-use util::write_projects_file;
+use util::in_workspace_with_projects_file;
 use util::Error;
 
 pub fn hash_set<I, T>(items: I) -> HashSet<T>
@@ -28,23 +24,6 @@ where
     T: Eq + Hash,
 {
     items.into_iter().collect()
-}
-
-fn in_workspace_with_projects_file<T, F>(projects_contents: &str, test: F) -> Result<T, Error>
-where
-    F: Fn(&Path, Workspace) -> Result<T, Error>,
-{
-    let tmpdir = TempDir::new("gws-test")?;
-    let workspace_dir = tmpdir.path().join("workspace");
-
-    create_dir_all(&workspace_dir)?;
-    write_projects_file(
-        workspace_dir.join(".projects.gws").as_path(),
-        projects_contents,
-    )?;
-
-    let workspace = read_workspace_file(workspace_dir.join(".projects.gws")).unwrap();
-    Ok(test(&workspace_dir, workspace)?)
 }
 
 #[test]
